+++ pkg/artifactory/resource/configuration/resource_artifactory_repository_layout.go
@@ -186,7 +186,7 @@ func (r *RepositoryLayoutResource) Create(ctx context.Context, req resource.Crea
 		return
 	}
 
-	err = SendConfigurationPatch(content, r.ProviderData)
+	err = SendConfigurationPatch(content, r.ProviderData.Client)
 	if err != nil {
 		utilfw.UnableToCreateResourceError(resp, fmt.Sprintf("failed to send PATCH request to Artifactory during Update: %s", err.Error()))
 		return
@@ -283,7 +283,7 @@ func (r *RepositoryLayoutResource) Update(ctx context.Context, req resource.Upda
 		return
 	}
 
-	err = SendConfigurationPatch(content, r.ProviderData)
+	err = SendConfigurationPatch(content, r.ProviderData.Client)
 	if err != nil {
 		utilfw.UnableToUpdateResourceError(resp, fmt.Sprintf("failed to send PATCH request to Artifactory during Update: %s", err.Error()))
 		return
@@ -328,7 +328,7 @@ repoLayouts:
   %s: ~
 `, matchedRepoLayout.Name)
 
-	err = SendConfigurationPatch([]byte(deleteConfig), r.ProviderData)
+	err = SendConfigurationPatch([]byte(deleteConfig), r.ProviderData.Client)
 	if err != nil {
 		utilfw.UnableToDeleteResourceError(resp, err.Error())
 		return